move view api cpp files
