+++ BUILD.gn
@@ -424,16 +424,16 @@ static_library("electron_lib") {
 
   if (enable_view_api) {
     sources += [
-      "atom/browser/api/atom_api_box_layout.cc",
-      "atom/browser/api/atom_api_box_layout.h",
-      "atom/browser/api/atom_api_button.cc",
-      "atom/browser/api/atom_api_button.h",
-      "atom/browser/api/atom_api_label_button.cc",
-      "atom/browser/api/atom_api_label_button.h",
-      "atom/browser/api/atom_api_layout_manager.cc",
-      "atom/browser/api/atom_api_layout_manager.h",
-      "atom/browser/api/atom_api_text_field.cc",
-      "atom/browser/api/atom_api_text_field.h",
+      "atom/browser/api/views/atom_api_box_layout.cc",
+      "atom/browser/api/views/atom_api_box_layout.h",
+      "atom/browser/api/views/atom_api_button.cc",
+      "atom/browser/api/views/atom_api_button.h",
+      "atom/browser/api/views/atom_api_label_button.cc",
+      "atom/browser/api/views/atom_api_label_button.h",
+      "atom/browser/api/views/atom_api_layout_manager.cc",
+      "atom/browser/api/views/atom_api_layout_manager.h",
+      "atom/browser/api/views/atom_api_text_field.cc",
+      "atom/browser/api/views/atom_api_text_field.h",
     ]
   }
