+++ atom/browser/api/views/atom_api_button.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#ifndef ATOM_BROWSER_API_ATOM_API_BUTTON_H_
-#define ATOM_BROWSER_API_ATOM_API_BUTTON_H_
+#ifndef ATOM_BROWSER_API_VIEWS_ATOM_API_BUTTON_H_
+#define ATOM_BROWSER_API_VIEWS_ATOM_API_BUTTON_H_
 
 #include "atom/browser/api/atom_api_view.h"
 #include "native_mate/handle.h"
@@ -35,4 +35,4 @@ class Button : public View, public views::ButtonListener {
 
 }  // namespace atom
 
-#endif  // ATOM_BROWSER_API_ATOM_API_BUTTON_H_
+#endif  // ATOM_BROWSER_API_VIEWS_ATOM_API_BUTTON_H_
