+++ atom/browser/api/views/atom_api_label_button.h
@@ -2,12 +2,12 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#ifndef ATOM_BROWSER_API_ATOM_API_LABEL_BUTTON_H_
-#define ATOM_BROWSER_API_ATOM_API_LABEL_BUTTON_H_
+#ifndef ATOM_BROWSER_API_VIEWS_ATOM_API_LABEL_BUTTON_H_
+#define ATOM_BROWSER_API_VIEWS_ATOM_API_LABEL_BUTTON_H_
 
 #include <string>
 
-#include "atom/browser/api/atom_api_button.h"
+#include "atom/browser/api/views/atom_api_button.h"
 
 namespace atom {
 
@@ -33,4 +33,4 @@ class LabelButton : public Button {
 
 }  // namespace atom
 
-#endif  // ATOM_BROWSER_API_ATOM_API_LABEL_BUTTON_H_
+#endif  // ATOM_BROWSER_API_VIEWS_ATOM_API_LABEL_BUTTON_H_
