+++ atom/browser/api/views/atom_api_layout_manager.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#ifndef ATOM_BROWSER_API_ATOM_API_LAYOUT_MANAGER_H_
-#define ATOM_BROWSER_API_ATOM_API_LAYOUT_MANAGER_H_
+#ifndef ATOM_BROWSER_API_VIEWS_ATOM_API_LAYOUT_MANAGER_H_
+#define ATOM_BROWSER_API_VIEWS_ATOM_API_LAYOUT_MANAGER_H_
 
 #include <memory>
 
@@ -41,4 +41,4 @@ class LayoutManager : public mate::TrackableObject<LayoutManager> {
 
 }  // namespace atom
 
-#endif  // ATOM_BROWSER_API_ATOM_API_LAYOUT_MANAGER_H_
+#endif  // ATOM_BROWSER_API_VIEWS_ATOM_API_LAYOUT_MANAGER_H_
