+++ atom/browser/api/views/atom_api_text_field.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#ifndef ATOM_BROWSER_API_ATOM_API_TEXT_FIELD_H_
-#define ATOM_BROWSER_API_ATOM_API_TEXT_FIELD_H_
+#ifndef ATOM_BROWSER_API_VIEWS_ATOM_API_TEXT_FIELD_H_
+#define ATOM_BROWSER_API_VIEWS_ATOM_API_TEXT_FIELD_H_
 
 #include "atom/browser/api/atom_api_view.h"
 #include "native_mate/handle.h"
@@ -39,4 +39,4 @@ class TextField : public View {
 
 }  // namespace atom
 
-#endif  // ATOM_BROWSER_API_ATOM_API_TEXT_FIELD_H_
+#endif  // ATOM_BROWSER_API_VIEWS_ATOM_API_TEXT_FIELD_H_
